+++ news.md
@@ -1,6 +1,7 @@
 # tealeaves (development version)
 
 * Added unit tests for parameter functions under `tests/test-parameter-functions.R`
+* Fixed bug with crossing parameters in `tleaves()` that was introduced with new `T_sky` function. This led to crossing all parameter values with all unique values of calculated `T_sky`, which was incorrect. Added unit tests to ensuring that crossing is done correctly under `tests/test-tleaves-crossing.R`
 * Added code coverage using [codecov](https://codecov.io/gh/cdmuir/tealeaves?branch=master)
 
 # tealeaves 1.0.3